Bags fixing

+++ eProiect/Controllers/HomeController.cs
@@ -227,6 +227,11 @@ namespace eProiect.Controllers
           [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult GetLoggedUserDisciplineTypes(int disciplineId)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             var loggedInUser = System.Web.HttpContext.Current.GetMySessionObject();
             var classTypes = _organizational.GetTypesByDisciplineForUser(disciplineId, loggedInUser.Id);
             
@@ -234,8 +239,14 @@ namespace eProiect.Controllers
         }
 
         [HttpGet]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult GetLoggedUserDisciplines()
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             var loggedInUser = System.Web.HttpContext.Current.GetMySessionObject();
             if (loggedInUser == null)
                 return Json(new List<Discipline>());
@@ -245,16 +256,29 @@ namespace eProiect.Controllers
         }
 
         [HttpGet]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult GetClassById(int classId)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             var resultClass=_organizational.GetClassById(classId);
             return Json(resultClass, JsonRequestBehavior.AllowGet);
         }
 
 
         [HttpPost]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult AddNewClass(ComposedClassInfo composedData, List<int> groupIds)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
+
             //check for negatives!!!!!!!!!!!!!!
             var properties = composedData.GetType().GetProperties();
             foreach(var property in properties)
@@ -329,8 +353,14 @@ namespace eProiect.Controllers
         }
         
         [HttpPost]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult GetOptionsByYear(int year)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             List<AcademicGroup> groupList = new List<AcademicGroup>();
 
             if (year == 0)
@@ -346,6 +376,11 @@ namespace eProiect.Controllers
           [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult GetFreeClassroomsByFloor(FreeClassroomsMinimalRequest requestData)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             /* var freeClassrooms = _organizational.GetFreeClassroomsByFloor(floor);
              return Json(freeClassrooms, JsonRequestBehavior.AllowGet);*/
             var freeClassrooms = _organizational.GetFreeClassroomsByFloorAndTime(
@@ -363,8 +398,14 @@ namespace eProiect.Controllers
         }
 
         [HttpPost]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult EditExistingClassroom(EditClassRequest data)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             //check equality.
             var UDClass = _organizational.GetClassById(data.ClassId);
             var oldClassParam = new EditClassRequest()
@@ -437,8 +478,14 @@ namespace eProiect.Controllers
         }
 
         [HttpPost]
+        [UserMode(UserRole.admin, UserRole.teacher)]
         public ActionResult DeleteClass(int id)
         {
+            SessionStatus();
+            if ((string)System.Web.HttpContext.Current.Session["LoginStatus"] != "login")
+            {
+                return RedirectToAction("Login", "Login");
+            }
             var result = _organizational.DeleteClassById(id);
             return Json(result);
         }
